bs-crawler/pyproject.toml
[tool.poetry]
name = "bs-crawler"
version = "0.0.1"
description = "beautifulsoup crawler"
authors = ["George Mazzeo <mazzeogeorge@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
crawlee = {version = "*", extras = ["beautifulsoup"]}
google-cloud-pubsub = "^2.27.1"
google-cloud-storage = "^2.18.2"
uuid6 = "^2024.7.10"
structlog-gcp = "^0.3.0"
structlog = "^24.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
